requirements.txt
# ─────────────────────────────────────────────
# Multi-Modal Clinical Decision Support System
# requirements.txt
# ─────────────────────────────────────────────

# Core ML
# Keep these as ranges so newer Python builds can resolve compatible wheels.
# For regulated deployments, freeze the resolved environment into a lockfile.
torch>=2.2,<3.0
torchvision>=0.17,<1.0
transformers>=4.40,<5.0
open_clip_torch>=2.24,<3.0
timm>=0.9,<2.0

# HuggingFace ecosystem
huggingface_hub>=0.22,<1.0
datasets>=2.19,<5.0
accelerate>=0.29,<2.0
safetensors>=0.4,<1.0

# Clinical NLP
biobert-embedding>=0.1.2,<0.2

# Explainability
grad-cam>=1.5,<2.0
captum>=0.7,<1.0

# Image processing
Pillow>=12.0.0
opencv-python-headless>=4.9,<5.0
albumentations>=1.4,<3.0
pydicom>=2.4,<4.0

# Data & utilities
numpy>=2.0.0
pandas>=2.2,<3.0
scikit-learn>=1.4,<2.0
scipy>=1.13,<2.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0
PyYAML>=6.0,<7.0

# API + serving
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0

# Streamlit demo
streamlit>=1.55.0
plotly>=6.0.0

# AWS
boto3>=1.34,<2.0
sagemaker>=2.219,<3.0
awscli>=1.32,<2.0

# Reporting
reportlab>=4.1,<5.0
jinja2>=3.1,<4.0

# Dev
python-dotenv>=1.0,<2.0
loguru>=0.7,<1.0
pytest>=8.2,<9.0
pytest-asyncio>=0.23,<2.0
black>=24.4,<26.0
